bundle-server: add region tracing
Using the logger in the 'bundleWebServer' struct, use 'TraceLogger.Region()'
to log web requests as a region.

Note that, unlike the 'git-bundle-server' regions, these regions will have
different session IDs than those of the program startup logs. This is
intentional, as the regions being added are scoped to the request, and are
not "parented" by the overarching web server session.

Signed-off-by: Victoria Dye <vdye@github.com>