e2e: add performance test
Add a test that captures timing measurements (in 'console.log()' printouts)
for cloning various large repositories. The tests are divided into two
example sets: one for tests that run (on modern hardware with a good
internet connection) in approx. 5 minutes (or less), and the other for tests
that take longer than that. The latter set is tagged with '@slow' and the
'default' profile is updated to skip them. Add an 'all' profile that runs
these slow performance tests alongside the "regular" tests.

Signed-off-by: Victoria Dye <vdye@github.com>