Makefile: add 'e2e-test' target
Add a target to the Makefile that installs the end-to-end test NPM
dependencies then runs the tests. By default, this will only run tests
matching the "default" profile of the tests (excluding performance tests).
The 'E2E_FLAGS' variable can be used to specify the profile(s) used by the
tests (e.g., '--all' to include the tests excluded by default).

Add documentation on how to run the end-to-end tests to 'README.md'.

Signed-off-by: Victoria Dye <vdye@github.com>