init: exit with error if route already exists
If 'init' tries to initialize a new route, but that route already exists on
disk (enabled or not), exit with an error. Without this explicit check, the
initialization will fail anyway (specifically, the bare repo clone), but the
call to 'RepositoryProvider.CreateRepository()' will enable the existing
route even if it was disabled. To avoid that unintended behavior and provide
clearer information to users, explicitly check for existing routes before
'CreateRepository()' and exit if the route already exists.

Signed-off-by: Victoria Dye <vdye@github.com>